+++ src/buildLibrary.ts
@@ -5,6 +5,9 @@ import path from 'path'
 import { buildSync } from 'esbuild'
 import { gzip } from 'zlib'
 import { exec } from 'child_process'
+import { Options } from '~lask'
+
+export type C = Options
 
 export function buildLibrary({
   name,
@@ -32,22 +35,31 @@ export function buildLibrary({
   const { log } = console
   try {
     // Build types
-    const dir = exec('tsc -p tsconfig.build.json --outFile dist/index.d.ts', () => {
+    const dir = exec(`tsc -p ${tsconfig} --outDir ${outdir}`, () => {
       log(`✔ ${name}: Built types`)
     })
 
     // Replace paths (if config / build config includes paths)
     dir.on('exit', () => {
-      const bcfg = require(tsconfig)
-      const cfg = require(path.join(process.cwd(), 'tsconfig.json'))
-      if (bcfg?.compilerOptions?.paths || cfg?.compilerOptions?.paths) {
-        exec('tsconfig-replace-paths --project tsconfig.build.json', () => {
+      const base_cfg = require(path.join(process.cwd(), 'tsconfig.json'))
+      let cfgWithPaths = ''
+      if (base_cfg?.compilerOptions?.paths) {
+        cfgWithPaths = 'tsconfig.json'
+      } else {
+        const cfg = require(tsconfig)
+        if (cfg?.compilerOptions?.paths) {
+          cfgWithPaths = tsconfig
+        }
+      }
+      if (cfgWithPaths) {
+        exec(`tsconfig-replace-paths --project ${cfgWithPaths}`, () => {
           log(`✔ ${name}: Resolved paths`)
         })
       }
     })
     // Build packages
     ;(Array.isArray(format) ? format : [format]).forEach((fmt) => {
+      const extension = fmt === 'esm' ? '.mjs' : '.js'
       const buildResult = buildSync({
         entryPoints,
         outdir,
@@ -56,7 +68,7 @@ export function buildLibrary({
         define,
         format: fmt,
         platform: isNode ? 'node' : 'neutral',
-        outExtension: { '.js': fmt === 'esm' ? '.mjs' : '.js' },
+        outExtension: { '.js': extension },
         bundle: true,
         metafile: true,
         minify: false,
@@ -67,7 +79,7 @@ export function buildLibrary({
 
       if (calculateSize) {
         // Calculate the size of the output esm module (standard and zipped)
-        fs.readFile('./dist/index.mjs', (_err, data) => {
+        fs.readFile(path.join(outdir, `index${extension}`), (_err, data) => {
           gzip(data, (_err, result) => {
             const size = Object.values(buildResult.metafile!.outputs).reduce(
               (acc, { bytes }) => acc + bytes,