+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 <!-- @format -->
 
+## 0.0.11
+
+- Fix paths handling.
+
+## 0.0.8
+
+- Fix paths detection.
+
+## 0.0.8
+
+- Improve paths detection.
+
+## 0.0.7
+
+- Fix type output.
+
+## 0.0.6
+
+- Fix dependencies.
+
+## 0.0.4
+
+- Simplify dependencies.
+
 ## 0.0.4
 
 - Add Github repo to package.json.